[WFLY-9339] Allow UndertowTransformersTestCase to work when WFCORE-3259 is fixed
